0-11

+++ 1-SimplyStorage.sol
@@ -1,7 +1,7 @@
 pragma solidity >=0.4.0 <0.6.0;
 
 contract SimpleStorage {
-    uint storedData;
+    uint256 storedData;
 
     function set(uint x) public {
         storedData = x;
